feat(tests): Add unit tests - Test the githubUsers model - Test githubUsersResponse model [Delivers #156185242]
