+++ app/src/androidTest/java/com/example/wagubibrian/github_connect/ExampleInstrumentedTest.java
@@ -1,6 +1,7 @@
 package com.example.wagubibrian.github_connect;
 
 import android.content.Context;
+
 import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
 
@@ -18,6 +19,7 @@ import static org.junit.Assert.*;
 public class ExampleInstrumentedTest {
     @Test
     public void useAppContext() {
+
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
