+++ app/src/androidTest/java/com/example/wagubibrian/github_connect/view/MainActivityTest.java
@@ -5,9 +5,11 @@ import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
 import com.example.wagubibrian.github_connect.R;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.Mock;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.Espresso.pressBack;
@@ -21,8 +23,11 @@ public class MainActivityTest {
     @Rule
     public ActivityTestRule<MainActivity> activity = new ActivityTestRule<MainActivity>(MainActivity.class);
 
+    @Mock
+    MainActivity activityTest;
+
     @Test
-    public void clickItemOnRecycler() throws Exception{
+    public void clickItemOnRecycler(){
         onView(withId(R.id.recycler_view))
                 .perform(RecyclerViewActions.scrollToPosition(0))
                 .perform(click());
@@ -39,8 +44,7 @@ public class MainActivityTest {
     }
 
     @Test
-    public void swipeToRefresh() throws Exception{
-        onView(withId(R.id.recycler_view)).perform(swipeDown());
+    public void swipeToRefresh(){
+        onView(withId(R.id.swipe_layout)).perform(swipeDown());
     }
-
 }
