+++ app/src/test/java/com/example/wagubibrian/github_connect/GithubUserResponseModelTest.java
@@ -9,48 +9,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.junit.Assert.*;
-
 /**
  * Example local unit test, which will execute on the development machine (host).
  *
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
-public class ExampleUnitTest {
-
-
-
-    @Test
-    public void additionIsCorrect() {
-        assertEquals(4, 2 + 2);
-    }
-
-
-    @Test
-    public void testGithubUserModel() {
-
-        String userName = "lubeja";
-        String profileImageUrl = "https://growth.search.com";
-        String email = "test@gmail.com";
-        String htmlUrl = "lubeja.com";
-        GithubUsers githubUser = new GithubUsers(userName, profileImageUrl, email,  htmlUrl);
-
-        assertEquals(userName ,githubUser.getUserName());
-        assertEquals(profileImageUrl, githubUser.getProfileImage());
-        assertEquals(githubUser.getEmail(), email);
-        assertEquals(htmlUrl, githubUser.getHtmlUrl());
-        assertEquals(githubUser.describeContents(), 0);
-
-
-        githubUser.setUserName("Changed");
-        githubUser.setProfileImage("https://profile.search.com");
-        githubUser.setEmail("changed@gmail.com");
-        githubUser.setHtmlUrl("changed.com");
-        assertNotEquals(userName ,githubUser.getUserName());
-        assertNotEquals(profileImageUrl, githubUser.getProfileImage());
-        assertNotEquals(githubUser.getEmail(), email);
-        assertNotEquals(htmlUrl, githubUser.getHtmlUrl());
-
-    }
+public class GithubUserResponseModelTest {
 
     @Test
     public void testGithubUserResponseModel() {
