+++ build.gradle
@@ -42,6 +42,7 @@ allprojects {
         okhttpVersion="2.7.0"
         mockwebserverVersion="2.7.0"
         espressoIntentsVersion="3.0.2"
+        mockitoCoreVersion="2.1.0"
     }
 }
